+++ CHANGELOG.md
@@ -44,7 +44,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 - Return proper error for unknown types in variable definitions
 - Ensure `_entities` maintains order of representations in result https://github.com/nuwave/lighthouse/pull/2286
 - Allow combining `@can` with non-default `@guard` https://github.com/nuwave/lighthouse/pull/2276
-- When a definition for a className cannot be found, instead of an empty classname being present in the exceptionMessage, the classname is now present.
+- Fix error message when failing to find class in namespace https://github.com/nuwave/lighthouse/pull/2342
 
 ### Added